refactor: move structs and macros to header
