+++ src/Primitives/arduino.cpp
@@ -139,170 +139,8 @@ int resolve_isr(int pin) {
 // Global index for installing primitives
 int prim_index = 0;
 
-/*
-   Private macros to install a primitive
-*/
-#define install_primitive(prim_name)                                       \
-    {                                                                      \
-        dbg_info("installing primitive number: %d  of %d with name: %s\n", \
-                 prim_index + 1, ALL_PRIMITIVES, #prim_name);              \
-        if (prim_index < ALL_PRIMITIVES) {                                 \
-            PrimitiveEntry *p = &primitives[prim_index++];                 \
-            p->name = #prim_name;                                          \
-            p->f = &(prim_name);                                           \
-        } else {                                                           \
-            FATAL("pim_index out of bounds");                              \
-        }                                                                  \
-    }
-
-#define def_prim(function_name, type) \
-    Type function_name##_type = type; \
-    bool function_name(Module *m)
-
-// TODO: use fp
-#define pop_args(n) m->sp -= n
-#define get_arg(m, arg) m->stack[(m)->sp - (arg)].value
-#define pushUInt32(arg) m->stack[++m->sp].value.uint32 = arg
-#define pushInt32(arg) m->stack[++m->sp].value.int32 = arg
-#define pushUInt64(arg)                 \
-    m->stack[++m->sp].value_type = I64; \
-    m->stack[m->sp].value.uint64 = arg
-#define arg0 get_arg(m, 0)
-#define arg1 get_arg(m, 1)
-#define arg2 get_arg(m, 2)
-#define arg3 get_arg(m, 3)
-#define arg4 get_arg(m, 4)
-#define arg5 get_arg(m, 5)
-#define arg6 get_arg(m, 6)
-#define arg7 get_arg(m, 7)
-#define arg8 get_arg(m, 8)
-#define arg9 get_arg(m, 9)
-
 // The primitive table
 PrimitiveEntry primitives[ALL_PRIMITIVES];
-
-//
-uint32_t param_arr_len0[0] = {};
-uint32_t param_I32_arr_len1[1] = {I32};
-uint32_t param_I32_arr_len2[2] = {I32, I32};
-uint32_t param_I32_arr_len3[3] = {I32, I32, I32};
-uint32_t param_I32_arr_len4[4] = {I32, I32, I32, I32};
-uint32_t param_I32_arr_len10[10] = {I32, I32, I32, I32, I32,
-                                    I32, I32, I32, I32, I32};
-
-uint32_t param_I64_arr_len1[1] = {I64};
-
-Type oneToNoneU32 = {
-    .form = FUNC,
-    .param_count = 1,
-    .params = param_I32_arr_len1,
-    .result_count = 0,
-    .results = nullptr,
-    .mask = 0x8001 /* 0x800 = no return ; 1 = I32*/
-};
-
-Type twoToNoneU32 = {
-    .form = FUNC,
-    .param_count = 2,
-    .params = param_I32_arr_len2,
-    .result_count = 0,
-    .results = nullptr,
-    .mask = 0x80011 /* 0x800 = no return ; 1 = I32; 1 = I32*/
-};
-
-Type threeToNoneU32 = {
-    .form = FUNC,
-    .param_count = 3,
-    .params = param_I32_arr_len3,
-    .result_count = 0,
-    .results = nullptr,
-    .mask = 0x800111 /* 0x800 = no return ; 1=I32; 1=I32; 1=I32*/
-};
-
-Type fourToNoneU32 = {
-    .form = FUNC,
-    .param_count = 4,
-    .params = param_I32_arr_len4,
-    .result_count = 0,
-    .results = nullptr,
-    .mask =
-        0x8001111 /* 0x800 = no return ; 1 = I32; 1 = I32; 1 = I32; 1 = I32*/
-};
-
-Type oneToOneU32 = {
-    .form = FUNC,
-    .param_count = 1,
-    .params = param_I32_arr_len1,
-    .result_count = 1,
-    .results = param_I32_arr_len1,
-    .mask = 0x80011 /* 0x8 1=I32 0=endRet ; 1=I32; 1=I32*/
-};
-
-Type oneToOneI32 = {
-    .form = FUNC,
-    .param_count = 1,
-    .params = param_I32_arr_len1,
-    .result_count = 1,
-    .results = param_I32_arr_len1,
-    .mask = 0x80011 /* 0x8 1=I32 0=endRet ; 1=I32; 1=I32*/
-};
-
-Type twoToOneU32 = {
-    .form = FUNC,
-    .param_count = 2,
-    .params = param_I32_arr_len2,
-    .result_count = 1,
-    .results = param_I32_arr_len1,
-    .mask = 0x81011 /* 0x8 1=I32 0=endRet ; 1=I32; 1=I32*/
-};
-
-Type threeToOneU32 = {
-    .form = FUNC,
-    .param_count = 3,
-    .params = param_I32_arr_len3,
-    .result_count = 1,
-    .results = param_I32_arr_len1,
-    .mask = 0x810111 /* 0x8 1=I32 0=endRet ; 1=I32; 1=I32; 1=I32*/
-};
-
-Type fourToOneU32 = {
-    .form = FUNC,
-    .param_count = 4,
-    .params = param_I32_arr_len4,
-    .result_count = 1,
-    .results = param_I32_arr_len1,
-    .mask = 0x8101111 /* 0x8 1=I32 0=endRet ; 1=I32; 1=I32; 1=I32; 1=I32*/
-};
-
-Type tenToOneU32 = {
-    .form = FUNC,
-    .param_count = 10,
-    .params = param_I32_arr_len10,
-    .result_count = 1,
-    .results = param_I32_arr_len1,
-    .mask = 0x8101111111111 /* 0x8 1=I32 0=endRet ; 10 params 1=I32*/
-};
-
-Type NoneToNoneU32 = {.form = FUNC,
-                      .param_count = 0,
-                      .params = nullptr,
-                      .result_count = 0,
-                      .results = nullptr,
-                      .mask = 0x80000};
-
-Type NoneToOneU32 = {.form = FUNC,
-                     .param_count = 0,
-                     .params = nullptr,
-                     .result_count = 1,
-                     .results = param_I32_arr_len1,
-                     .mask = 0x81000};
-
-Type NoneToOneU64 = {.form = FUNC,
-                     .param_count = 0,
-                     .params = nullptr,
-                     .result_count = 1,
-                     .results = param_I64_arr_len1,
-                     .mask = 0x82000};
 
 // Util function declarations
